Complete struct-of-arrays migration (Performance Improvement #6)
Successfully migrated from array-of-arrays to struct-of-arrays data layout:

- Converted 100+ array access patterns across outputControl.c and metropolisJoint.c
- Used systematic sed-driven approach following compiler errors
- All array syntax now uses compatibility macros: r[nf][i][coord] -> r(nf,i,coord)
- Updated Makefile to include polymer_data_structures.c in all targets
- Zero compilation errors - migration foundation complete
- Documented working conversion strategy in MIGRATION_GUIDE.md

Foundation now ready for performance optimizations (bulk memcpy, better cache locality).

+++ src/PolymerCode/polymer_data_structures.c
@@ -10,7 +10,7 @@
 simulation_state_t g_sim_state = {0};
 #endif
 
-void allocate_simulation_state(simulation_state_t *state, int NFil, const int *N) {
+void allocate_simulation_state(simulation_state_t *state, int NFil, const long *N) {
     if (state->allocated) {
         printf("Warning: Simulation state already allocated. Freeing first.\n");
         free_simulation_state(state);
